maybe skip Terminus installation
Introduces a new `skip_terminus_install` boolean input. When set to `true`, the action will bypass the check and installation steps for Terminus.

This is useful for workflows using custom Docker images where Terminus is pre-installed, or when Terminus is installed in an earlier step. This change improves workflow efficiency and flexibility for advanced use cases.

By default, the action maintains its original behavior of checking for and installing Terminus if it's not found. But if the input is set to true or if the Terminus check runs and Terminus is already installed, the workflow does not attempt to install Terminus again.

Closes: #71