Release management updates (#98)
* break the release step into two
since we want to require PRs  and PR approvals before merging, we need a PR for readme updates. This necessitates having a two-step prepare-and-release process. This commit breaks the release process in half and creates a PR with the docs/readme changes to merge immediately before a release. additionally, the release workflow is updated to ensure that the prepare step was run first.

* update contributing to match the new workflow